Codebase modifications and standardizations: - Move SQL file to new /resources directory - Change indentation to 2 spaces - Remove debug print calls (can we use a log system for loggable messages) - Break out core pages and concepts in to their own modules - New init file is Spark.py, also importable - Move JWT to APP section of env vars
