+++ README.md
@@ -5,8 +5,8 @@
 * Clone and CD in: `git clone git@github.com:Yeungdb/Canna.git && cd Canna`
 * Install deps: `pip install Flask Flask-Login psycopg2 requests urllib3 twilio wit`
 * Set environment variable: `export CANNAKEY = 'YOURKEY'`
-* Create DB and set up structure: `createdb Canna && psql -h localhost -d Canna -a -q -f $PWD/Canna.sql`
-* Start it up: `python DispensaryPage.py`
+* Create DB and set up structure: `createdb Canna && psql -h localhost -d Canna -a -q -f $PWD/resources/Canna.sql`
+* Start it up: `python Spark.py`
 
 You can reset the whole DB using `dropdb Canna && createdb Canna`
